+++ lib/src/gradient_circle_percentage.dart
@@ -30,8 +30,11 @@ class GradientCirclePercentage extends StatefulWidget {
   /// This parameter is the circle color behind of the percentage
   final Color backgroundColor;
 
+  /// This parameter is a custom center text with default value null (if it null then show the percentage value)
+  final String? centerText;
+
   /// This parameter is the text style of the label for the percentage text
-  final TextStyle textStyle;
+  final TextStyle centerTextStyle;
 
   const GradientCirclePercentage({
     super.key,
@@ -44,7 +47,8 @@ class GradientCirclePercentage extends StatefulWidget {
     this.bottomColor = Colors.black,
     this.topColor = Colors.white,
     this.backgroundColor = Colors.white,
-    this.textStyle = const TextStyle(color: Colors.black),
+    this.centerText,
+    this.centerTextStyle = const TextStyle(color: Colors.black),
   })  : assert(currentPercentage <= maxPercentage),
         assert(currentPercentage >= 0),
         assert(duration >= 0);
@@ -84,8 +88,9 @@ class _GradientCirclePercentageState extends State<GradientCirclePercentage> {
             ),
             child: Center(
               child: Text(
-                (value * widget.maxPercentage).toInt().toString(),
-                style: widget.textStyle,
+                widget.centerText ??
+                    (value * widget.maxPercentage).toInt().toString(),
+                style: widget.centerTextStyle,
               ),
             ),
           ),